Fixed bugs in Mac OS X compilation

+++ joint_genotype_block_record.cpp
@@ -669,7 +669,7 @@ void JointGenotypeBlockRecord::process_read(AugmentedBAMRecord& as, int32_t samp
 		allele = 1;
 	      }
 	      else {
-		q = abs(len-aug_ref[i].size())*30;
+		q = (int32_t)abs((double)(len-aug_ref[i].size()))*30;
 		allele = -1;
 	      }
 	      
@@ -692,7 +692,7 @@ void JointGenotypeBlockRecord::process_read(AugmentedBAMRecord& as, int32_t samp
 		allele = 1;
 	      }
 	      else {
-		q = abs(len-aug_ref[i].size())*30;
+		q = (int32_t)abs((double)(len-aug_ref[i].size()))*30;
 		allele = -1;
 	      }
 	      cycle = strand == 'F' ? (rpos0+1) : (rlen - rpos0);
